+++ src/ni_measurement_plugin_packager/__init__.py
@@ -10,17 +10,6 @@ import click
 from nisystemlink_feeds_manager.clients.core import ApiException
 from pydantic import ValidationError
 
-from ni_measurement_plugin_packager.constants import (
-    CliInterface,
-    InteractiveModeMessages,
-    NonInteractiveModeMessages,
-    UserMessages,
-)
-from ni_measurement_plugin_packager.models import (
-    CliInputs,
-    SystemLinkConfig,
-    UploadPackageInfo,
-)
 from ni_measurement_plugin_packager._support._helpers import (
     build_package,
     get_publish_package_client,
@@ -37,6 +26,17 @@ from ni_measurement_plugin_packager._support._logger import (
 )
 from ni_measurement_plugin_packager._support._non_interactive_mode import (
     publish_packages_in_non_interactive_mode,
+)
+from ni_measurement_plugin_packager.constants import (
+    CliInterface,
+    InteractiveModeMessages,
+    NonInteractiveModeMessages,
+    UserMessages,
+)
+from ni_measurement_plugin_packager.models import (
+    CliInputs,
+    SystemLinkConfig,
+    UploadPackageInfo,
 )
 
 CONTEXT_SETTINGS = {"help_option_names": ["-h", "--help"]}
