+++ src/ni_measurement_plugin_packager/_support/_create_files.py
@@ -2,11 +2,9 @@
 
 import platform
 import shutil
-import sys
 from pathlib import Path
 
-import winreg
-
+from ni_measurement_plugin_packager._support import _get_nipath
 from ni_measurement_plugin_packager.constants import (
     ControlFile,
     FileNames,
@@ -31,21 +29,6 @@ ignore_dirs = [
 
 def _get_measurement_services_path(measurement_name: str) -> Path:
     return _get_nipath("NIPUBAPPDATADIR") / "Plug-Ins" / "Measurements" / measurement_name
-
-
-def _get_nipath(name: str) -> Path:
-    if sys.platform == "win32":
-        access: int = winreg.KEY_READ
-        if "64" in name:
-            access |= winreg.KEY_WOW64_64KEY
-        with winreg.OpenKey(
-            winreg.HKEY_LOCAL_MACHINE,
-            r"SOFTWARE\National Instruments\Common\Installer",
-            access=access,
-        ) as key:
-            value, type = winreg.QueryValueEx(key, name)
-            assert type == winreg.REG_SZ  # nosec: B101
-            return Path(value)
 
 
 def _get_system_type() -> str:
