+++ src/ni_measurement_plugin_packager/_support/_non_interactive_mode.py
@@ -6,13 +6,13 @@ from typing import List
 
 from nisystemlink_feeds_manager.main import PublishPackagesToSystemLink
 
-from ni_measurement_plugin_packager.constants import UserMessages
-from ni_measurement_plugin_packager.models import InvalidInputError, UploadPackageInfo
 from ni_measurement_plugin_packager._support._helpers import (
     get_folders,
     publish_packages,
     validate_selected_plugins,
 )
+from ni_measurement_plugin_packager.constants import UserMessages
+from ni_measurement_plugin_packager.models import InvalidInputError, UploadPackageInfo
 
 
 def publish_packages_in_non_interactive_mode(
