+++ src/ni_measurement_plugin_packager/_support/_user_inputs.py
@@ -4,6 +4,11 @@ from logging import Logger
 from pathlib import Path
 from typing import Dict, List, Optional, Tuple
 
+from ni_measurement_plugin_packager._support._helpers import (
+    display_available_measurements,
+    get_folders,
+    get_measurement_plugins,
+)
 from ni_measurement_plugin_packager.constants import (
     YES,
     InteractiveModeMessages,
@@ -13,11 +18,6 @@ from ni_measurement_plugin_packager.models import (
     InvalidInputError,
     SystemLinkConfig,
     UploadPackageInfo,
-)
-from ni_measurement_plugin_packager._support._helpers import (
-    display_available_measurements,
-    get_folders,
-    get_measurement_plugins,
 )
 
 ALL_MEAS = "."
